rename struct for json bytes
